requirements.txt
ultralytics>=8.4.19
onnxruntime>=1.24.2
opencv-python>=4.13.0.92
numpy>=2.4.2
pytest>=9.0.2
PyYAML>=6.0.3
customtkinter>=5.2.2
Pillow>=12.1.1
